# Memoize FQL template regex