--- a/query_test.go
+++ b/query_test.go
@@ -88,6 +88,6 @@ func TestFQL(t *testing.T) {
 
 func BenchmarkFQL(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		FQL(`${arg0}.length`, map[string]any{"arg0": "foo"})
+		_, _ = FQL(`${arg0}.length`, map[string]any{"arg0": "foo"})
 	}
 }