Ticket(s): BT-###, FE-###,...

## Problem

Benchmarks suggest that the majority of the query building time is spent compiling its template regex.

## Solution

Memoize compiled regex.

## Result

Before path:

```
go test -bench=BenchmarkFQL
goos: darwin
goarch: amd64
pkg: github.com/fauna/fauna-go
cpu: Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
BenchmarkFQL-8            182796              6310 ns/op
PASS
ok      github.com/fauna/fauna-go       4.690s
```

After the patch:

```
go test -bench=BenchmarkFQL
goos: darwin
goarch: amd64
pkg: github.com/fauna/fauna-go
cpu: Intel(R) Core(TM) i7-7820HQ CPU @ 2.90GHz
BenchmarkFQL-8            767666              1416 ns/op
PASS
ok      github.com/fauna/fauna-go       4.148s
```

## Testing

Existing unit tests apply. Benchmark test included.

----
By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.